refactor ui trace graph merge
